refactor: merge MarketplaceUninstaller into MarketplaceInstaller
